A starting point
A decision was made to have all peripherals as subtypes of
T_PERIPHERALS. Thus the bus can speak to all and just gets ignored by
most of the listeners. This can also be used to configure peripherals by
writing to a location that is read only.
Unfortunately the program doesn't compile yet. have to figure out what
is wrong with Bus.